# docs/api.yml
openapi: "3.0.2"
info:
  title: Scholarship API
  version: "1.0"
servers:
  - url: http://localhost:8080
  - url: https://virtserver.swaggerhub.com/AmourRamanantsiresy/Scholarship/1.0.0
paths:
...
  /language:
    get:
      tags:
        - "Language"
      operationId: getAllLanguages
      parameters:
        - $ref: '#/components/parameters/PSize'
        - $ref: '#/components/parameters/PPage'
        - $ref: '#/components/parameters/POrder'
      responses:
        "200":
          description: "Success"
          content:
            application/json:
              schema:
                type: object
                properties:
                  last_page:
                    type: number
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/IdName'
    put:
      tags:
        - Language
      operationId: createOrUpdateLanguage
      description: Update an existing language if you specify the id else create new language
      requestBody:
        content:
          application/json:
            schema: 
              type: array
              items:
                $ref: '#/components/schemas/IdName'
      responses:
        "200":
          description: "Success"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/IdName'
  /language/search:
    get:
      tags:
        - "Language"
      operationId: searchLanguages
      parameters:
        - $ref: '#/components/parameters/PSearch'
        - $ref: '#/components/parameters/PSize'
        - $ref: '#/components/parameters/PPage'
        - $ref: '#/components/parameters/POrder'
        
      responses:
        "200":
          description: "Success"
          content:
            application/json:
              schema:
                type: object
                properties:
                  last_page:
                    type: number
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/IdName'
...
  /candidate:
    get:
      tags:
        - "Candidate"
      operationId: getAllCandidates
      parameters:
        - $ref: '#/components/parameters/PSize'
        - $ref: '#/components/parameters/PPage'
        - $ref: '#/components/parameters/POrder'
        - $ref: '#/components/parameters/PCandidate'
      responses:
        "200":
          description: "Success"
          content:
            application/json:
              schema:
                type: object
                properties:
                  last_page:
                    type: number
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/RestCandidate'
    put:
      tags:
        - Candidate
      operationId: createOrUpdateCandidate
      description: Update an existing candidate if you specify the id else create new candidate
      requestBody:
        content:
          application/json:
            schema: 
              type: array
              items:
                $ref: '#/components/schemas/CreateCandidate'
      responses:
        "200":
          description: "Success"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RestCandidate'
  /candidate/search:
    get:
      tags:
        - "Candidate"
      operationId: searchCandidates
      parameters:
        - $ref: '#/components/parameters/PSearch'
        - $ref: '#/components/parameters/PSize'
        - $ref: '#/components/parameters/PPage'
        - $ref: '#/components/parameters/PCandidate'
      responses:
        "200":
          description: "Success"
          content:
            application/json:
              schema:
                type: object
                properties:
                  last_page:
                    type: number
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/RestCandidate'
  /scholarship:
    get:
      tags:
        - "Scholarship"
      operationId: getAllScholarships
      parameters:
        - $ref: '#/components/parameters/PSize'
        - $ref: '#/components/parameters/PPage'
        - $ref: '#/components/parameters/POrder'
        - $ref: '#/components/parameters/PScholarship'

      responses:
        "200":
          description: "Success"
          content:
            application/json:
              schema:
                type: object
                properties:
                  last_page:
                    type: number
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/RestScholarship'
    put:
      tags:
        - Scholarship
      operationId: createOrUpdateScholarship
      description: Update an existing scholarship if you specify the id else create new scholarship
      requestBody:
        content:
          application/json:
            schema: 
              type: array
              items:
                $ref: '#/components/schemas/CreateScholarship'
      responses:
        "200":
          description: "Success"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RestScholarship'
  /scholarship/search:
    get:
      tags:
        - "Scholarship"
      operationId: searchScholarships
      parameters:
        - $ref: '#/components/parameters/PSearch'
        - $ref: '#/components/parameters/PSize'
        - $ref: '#/components/parameters/PPage'
        - $ref: '#/components/parameters/PScholarshipSearch'
      responses:
        "200":
          description: "Success"
          content:
            application/json:
              schema:
                type: object
                properties:
                  last_page:
                    type: number
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/RestScholarship'
  /scholarship/search/amount:
    get:
      tags:
        - "Scholarship"
      operationId: searchScholarshipsByAmount
      description: find all scholarship that have a specified amount
      parameters:
        - name: amount
          in: query
          required: true
          schema:
            type: integer
        - $ref: '#/components/parameters/PSize'
        - $ref: '#/components/parameters/PPage'
      responses:
        "200":
          description: "Success"
          content:
            application/json:
              schema:
                type: object
                properties:
                  last_page:
                    type: number
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/RestScholarship'
  /scholarship/search/duration:
    get:
      tags:
        - "Scholarship"
      operationId: searchScholarshipsByDuration
      description: search all scholarship that have duration between two specified durations
      parameters:
        - name: duration1
          in: query
          required: true
          schema:
            type: integer
        - name: duration2
          in: query
          required: true
          schema:
            type: integer
        - $ref: '#/components/parameters/PSize'
        - $ref: '#/components/parameters/PPage'
      responses:
        "200":
          description: "Success"
          content:
            application/json:
              schema:
                type: object
                properties:
                  last_page:
                    type: number
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/RestScholarship'
  /scholarship/search/creation:
    get:
      tags:
        - "Scholarship"
      operationId: searchScholarshipsByCreation
      description: search all scholarships created between two specified dates
      parameters:
        - name: beginning
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: ending
          in: query
          required: true
          schema:
            type: string
            format: date
        - $ref: '#/components/parameters/PSize'
        - $ref: '#/components/parameters/PPage'
      responses:
        "200":
          description: "Success"
          content:
            application/json:
              schema:
                type: object
                properties:
                  last_page:
                    type: number
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/RestScholarship'
  /application/candidate/{id}:
    get:
      tags:
        - "Application"
      operationId: getAllApplicationsByCandidate
      parameters:
        - $ref: '#/components/parameters/PSize'
        - $ref: '#/components/parameters/PPage'
        - name: id
          in: path
          description: The candidate id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: "Success"
          content:
            application/json:
              schema:
                type: object
                properties:
                  last_page:
                    type: number
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        status:
                          enum : [PENDING, ACCEPTED]
                        scholarship: 
                          $ref: '#/components/schemas/RestScholarship'
  /application/scholarship/{id}/candidate:
    get:
      tags:
        - "Application"
      operationId: getAllCandidateByScholarship
      description: Get all candidate who applied for a specified scholarship
      parameters:
        - $ref: '#/components/parameters/PSize'
        - $ref: '#/components/parameters/PPage'
        - name: id
          in: path
          description: The scholarship id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: "Success"
          content:
            application/json:
              schema:
                type: object
                properties:
                  last_page:
                    type: number
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        status:
                          enum : [PENDING, ACCEPTED]
                        candidate:
                          $ref: '#/components/schemas/RestCandidateMini'
...
components:
  schemas:
...
    IdName:
      type: object
      properties:
        id: 
          type: integer
        name: 
          type: string
    Amount:
      type: object
      properties:
        id: 
          type: integer
        amount: 
          type: integer
...
    Candidate:
      allOf:
        - $ref: '#/components/schemas/RestCandidateMini'
      type: object
      properties:
        email:
          type: string
          format: email
        school_origin:
          type: string
        about: 
          type: string
        phone_number:
          pattern: '^\+261 \d{2} \d{2} \d{3} \d{2}$'
          type: string
        birth_date:
          type: string
          format: date
    Scholarship:
      type: object
      properties:
        id: 
          type: integer
        description:
          type: string
        duration:
          type: integer
          description: The duration of the study in year
        registration_deadline:
          type: string
          format: date
        status:
          type: string
          enum: [AVAILABLE, UNAVAILABLE]

  # schemas when create or update
...
    CreateUniversity:
      properties:
        country:
          type: array
          items:
            type: object
            properties:
              id:
                type: integer
      allOf:
        - $ref: '#/components/schemas/IdName'
    CreateCandidate:
      allOf:
        - $ref: '#/components/schemas/Candidate'
      type: object
      properties:
        country:
          type: array
          items:
            type: object
            properties:
              id:
                type: integer
        study_level: 
          type: array
          items:
            type: object
            properties:
              id:
                type: integer
        credentials: 
          type: array
          items:
            type: object
            properties:
              id:
                type: integer
    CreateScholarship:
      allOf:
        - $ref: '#/components/schemas/Scholarship'
      type: object
      properties:
        university:
          $ref: '#/components/schemas/CreateUniversity'
        faculty:
          $ref: '#/components/schemas/IdName'
        amount:
          $ref: '#/components/schemas/Amount'
  
  # schemas when get
...
    RestCandidate:
      allOf:
        - $ref: '#/components/schemas/Candidate'
      type: object
      properties:
        country:
          type: string
        languages:
          type: array
          items:
            type: string
            example: french
        study_level: 
          type: string
    RestCandidateMini:
      type: object
      properties:
        id:
          type: integer
        last_name: 
          type: string
        first_name: 
          type: string
    RestScholarship:
      allOf:
        - $ref: '#/components/schemas/Scholarship'
      type: object
      properties:
        faculty:
          type: string
        amount:
          type: string
        university:
          type: string
        country:
          type: string
        language:
          type: array
          items:
            type: string
            example: America
...
  parameters:
    PSize:
      name: size
      in: query
      required: false
      schema:
        type: integer
    PPage:
      name: page
      in: query
      required: false
      schema:
        type: integer
    POrder:
      name: order
      in: query
      required: false
      schema:
        enum: [ASC, DESC]
    PSearch:
      name: value
      in: query
      required: true
      schema:
        type: string
...
    PCandidate:
      name: by
      in: query
      required: false
      schema:
        type: string
        enum: [OLD, SCHOOL_ORIGIN, STUDY_LEVEL, FIRST_NAME, LAST_NAME]
    PScholarship: 
      name: by
      in: query
      required: false
      schema:
        type: string
        enum: [LANGUAGE, COUNTRY, DEADLINE, FACULTY, AMOUNT, UNIVERSITY, DURATION]
    PScholarshipSearch:
      name: by
      in: query
      required: false
      schema:
        type: string
        enum: [LANGUAGE, COUNTRY, FACULTY, UNIVERSITY]